Fix mc create_buckets script
There were problems stemming from the fact that the minio/mc image
doesn't have bash installed, so the script wasn't running correctly.

Besides, associative arrays don't work with the sh version that the
image has, so I had to change the structure of the script a little (and
it ended up better, in my opinion).